add typeinfomation unit-test
